New feat.
We rename the .cif files of the examples and add a new read_structure
function to directly read those structure into StructureData.

We refactor the USFWorkChain and GSFEWorkChain using PwBaseWorkChain
instead of directly call PwCalculation for more robustness. But they are
still to be checked.

We added builder_from_protocol function for more convenient construction
of the builder.

We remove the thermo_pw workchains and it's a separate package
aiida-thermo_pw

We transition the construction of the supercell by ase built-in
functions instead of implement them by ourselves.